+++ R/common-params.R
@@ -1,8 +1,13 @@
 
 
 #' @name common-params
-#' @param id Unique ID string. By default, name of the calling package.
-#' @param path Directory to store lockfiles.
+#' @param id Unique ID string. If this is unset, the name of the calling
+#'   package will be used. Since onetime 0.2.0, not setting `id` is
+#'   deprecated.
+#' @param path Directory to store lockfiles. The default uses a unique
+#'   directory corresponding to the calling package, beneath 
+#'   [rappdirs::user_config_dir()]. Normally you should leave this as the 
+#'   default.
 #' @param expiry [difftime()] or e.g. [lubridate::duration()] object.
 #'   After this length of time, code will be run again.
 #' @param message Message to display to the user.
@@ -14,7 +19,8 @@
 #'   simply presses return.
 #' @param without_permission Character string. What to do if the user hasn't
 #'   given permission to store files? `"warn"` runs the action with an extra
-#'   warning; `"run"` runs the action; `"pass"` does nothing and returns the
-#'   default; `"stop"` throws an error; `"ask"` asks for permission, after
-#'   running the action but before recording it on disk.
+#'   warning; `"run"` runs the action with no warning; `"pass"` does nothing
+#'   and returns the default; `"stop"` throws an error; `"ask"` asks for
+#'   permission using [check_ok_to_store()], and returns the default if it is
+#'   not granted.
 NULL