+++ R/messages.R
@@ -6,52 +6,54 @@
 #' @param ... Passed to [warning()], [message()] or [packageStartupMessage()].
 #' @inherit common-params
 #'
-#' @return `TRUE` if the message/warning was shown, `FALSE` otherwise.
+#' @return Invisible `TRUE` if the message/warning was shown, invisible
+#'   `FALSE` otherwise.
 #'
 #' @export
 #'
 #' @seealso [onetime_do()]
 #'
-#' @examples
+#' @doctest
 #' oo <- options(onetime.dir = tempdir(check = TRUE))
 #' id <- sample(10000L, 1)
 #'
 #' for (n in 1:3) {
+#' @expect warning(regexp = if (n == 1L) "once" else NA)
 #'   onetime_warning("will be shown once", id = id)
 #' }
 #'
 #' onetime_reset(id = id)
 #' options(oo)
 onetime_warning <- function(...,
-        id     = calling_package(),
+        id     = deprecate_calling_package(),
         path   = default_lockfile_dir(),
         expiry = NULL,
         without_permission = "warn"
       ) {
   ret_val <- onetime_do(warning(..., call. = FALSE), id = id, path = path, expiry = expiry,
                   default = FALSE, without_permission = without_permission)
-  return(! isFALSE(ret_val))
+  return(invisible(! isFALSE(ret_val)))
 }
 
 
 #' @rdname onetime_warning
 #' @export
 onetime_message <- function (...,
-        id     = calling_package(),
+        id     = deprecate_calling_package(),
         path   = default_lockfile_dir(),
         expiry = NULL,
         without_permission = "warn"
       ) {
   ret_val <- onetime_do(message(...),  id = id, path = path, expiry = expiry,
                         default = FALSE, without_permission = without_permission)
-  return(! isFALSE(ret_val))
+  return(invisible(! isFALSE(ret_val)))
 }
 
 
 #' @rdname onetime_warning
 #' @export
 onetime_startup_message <- function (...,
-        id     = calling_package(),
+        id     = deprecate_calling_package(),
         path   = default_lockfile_dir(),
         expiry = NULL,
         without_permission = "warn"
@@ -59,15 +61,14 @@ onetime_startup_message <- function (...,
   ret_val <- onetime_do(packageStartupMessage(...), id = id, path = path,
                         expiry = expiry, default = FALSE,
                         without_permission = without_permission)
-  return(! isFALSE(ret_val))
+  return(invisible(! isFALSE(ret_val)))
 }
 
 
 #' Print a message, and ask for confirmation to hide it in future
 #'
 #' This uses [readline()] to ask the user if the message should
-#' be shown again in future. In a non-interactive session, it does
-#' nothing.
+#' be shown again in future.
 #'
 #' By default, the message will be hidden if the user answers
 #' "n", "No", or "N", or just presses return to the prompt question.
@@ -81,53 +82,85 @@ onetime_startup_message <- function (...,
 #'
 #'
 #' @inherit common-params
+#' @param ... Passed to [message()].
 #' @param require_permission Logical. Ask permission to store files on the user's
 #'  computer, if this hasn't been granted? Setting this to `FALSE`
 #'  overrides `without_permission`.
+#' @param noninteractive String. Additional message to send in non-interactive
+#'  sessions. Set to `NULL` to do nothing in non-interactive sessions. The
+#'  default tells the user how to manually mark the message as done.
+#' @param message Deprecated. Use unnamed arguments `...` instead.
 #'
-#' @return `NULL` if the message was not shown (shown already or non-interactive
-#'   session). `TRUE` if the user confirmed (i.e. asked to hide the message).
-#'   `FALSE` if the message was shown but the user did not confirm. Note that by
-#'   default, `TRUE` is returned when the user answers "no" to "Show this message
-#'   again?"
+#' @return
+#' * `NULL` if the message was not shown (shown already or non-interactive
+#'   session and `noninteractive` was `NULL`).
+#' * `TRUE` if the user confirmed, i.e. chose to hide the message.
+#' * `FALSE` if the message was shown but the user did not confirm (did not
+#'   choose to hide the message, or non-interactive session and `noninteractive`
+#'   was not `NULL`).
+#'
+#' Results are returned invisibly.
+#'
+#' Note that by default, `TRUE` is returned when the user answers "no" to
+#' "Show this message again?" and `FALSE` is returned when the user answers
+#' "yes".
 #'
 #' @export
 #'
-#' @examples
+#' @doctest
 #' oo <- options(onetime.dir = tempdir(check = TRUE))
 #' id <- sample(10000L, 1L)
-#'
+#' @expect message("A message")
 #' onetime_message_confirm("A message to show one or more times", id = id)
 #'
 #' onetime_reset(id = id)
 #' options(oo)
-onetime_message_confirm <- function (message,
-  id              = calling_package(),
-  path            = default_lockfile_dir(),
-  expiry          = NULL,
-  confirm_prompt  = "Show this message again? [yN] ",
-  confirm_answers = c("N", "n", "No", "no"),
-  default_answer  = "N",
-  require_permission  = FALSE,
-  without_permission = "warn"
+onetime_message_confirm <- function (
+  ...,
+  id                 = deprecate_calling_package(),
+  path               = default_lockfile_dir(),
+  expiry             = NULL,
+  confirm_prompt     = "Show this message again? [yN] ",
+  confirm_answers    = c("N", "n", "No", "no"),
+  default_answer     = "N",
+  require_permission = FALSE,
+  without_permission = "warn",
+  noninteractive     = paste0(
+    "To hide this message in future, run:\n",
+    "  onetime::onetime_mark_as_done(id = \"", id, "\")"),
+  message            = .Deprecated()
 ) {
-  if (! my_interactive()) return(NULL)
+  dots <- list(...)
+  if (! missing(message)) {
+    .Deprecated("onetime_message_confirm(message = ...)", "onetime_message_confirm(...)")
+    dots <- list(message)
+  }
+
+  if (! my_interactive()) {
+    if (is.null(noninteractive)) {
+      return(invisible(NULL)) # message not shown
+    } else {
+      dots[length(dots) + 1:2] <- c("\n", noninteractive)
+      do.call(base::message, dots)
+      return(invisible(FALSE)) # user did not confirm
+    }
+  }
 
   confirmation <- expression({
-    message(message)
+    do.call(base::message, dots)
     answer <- my_readline(confirm_prompt)
     answer
   })
 
   answer <- do_onetime_do(confirmation, id = id, path = path, expiry = expiry,
                        without_permission = without_permission,
-                       require_permission = require_permission)
-  if (is.null(answer)) return(NULL)
+                       require_permission = require_permission, invisibly = TRUE)
+  if (is.null(answer)) return(invisible(NULL))
 
   if (answer == "") answer <- default_answer
   if (! answer %in% confirm_answers) {
     onetime_reset(id, path)
   }
 
-  return(answer %in% confirm_answers)
+  return(invisible(answer %in% confirm_answers))
 }