+++ R/rlang.R
@@ -6,13 +6,14 @@
 #' @param ... Passed to [rlang::warn()] or [rlang::inform()].
 #' @inherit common-params
 #'
-#' @return `TRUE` if the message/warning was shown, `FALSE` otherwise.
+#' @return Invisibly: `TRUE` if the message/warning was shown, `FALSE` otherwise.
 #'
-#' @examples
+#' @doctest
 #' oo <- options(onetime.dir = tempdir(check = TRUE))
 #' id <- sample(10000L, 1)
 #'
 #' for (n in 1:3) {
+#' @expect warning(regexp = if (n == 1L) "rlang" else NA)
 #'   onetime_rlang_warn(c("rlang-style warning", i = "Extra info"), id = id)
 #' }
 #'
@@ -25,7 +26,7 @@ NULL
 #' @rdname onetime-rlang
 #' @export
 onetime_rlang_warn <- function (...,
-        id     = calling_package(),
+        id     = deprecate_calling_package(),
         path   = default_lockfile_dir(),
         expiry = NULL,
         without_permission = "warn"
@@ -36,14 +37,14 @@ onetime_rlang_warn <- function (...,
                     id = id, path = path, expiry = expiry,
                     without_permission = without_permission
                    )
-  return(! is.null(res))
+  return(invisible(! is.null(res)))
 }
 
 
 #' @rdname onetime-rlang
 #' @export
 onetime_rlang_inform <- function (...,
-        id     = calling_package(),
+        id     = deprecate_calling_package(),
         path   = default_lockfile_dir(),
         expiry = NULL,
         without_permission = "warn"
@@ -56,5 +57,5 @@ onetime_rlang_inform <- function (...,
                     default = "not null",
                     without_permission = without_permission
                    )
-  return(is.null(res)) # ... and return TRUE if you got NULL
+  return(invisible(is.null(res))) # ... and return TRUE if you got NULL
 }